# api/openapi.yaml
openapi: 3.0.3
info:
  title: AntiNAT Controller API
  version: 1.0.0-beta
  description: >
    Frozen AntiNAT v1.0-beta management API. All resource modifications use
    ETag/If-Match; missing precondition returns 428, revision mismatch 412,
    and port/state conflict 409. Creates accept a durable Idempotency-Key.
    Error responses use the registry in docs/error-codes.md.
servers:
  - url: /
...
paths:
...
  /api/v1/auth/login:
    post:
      tags: [auth]
      operationId: login
      summary: Authenticate an administrator
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/LoginRequest" }
      responses:
        "200":
          description: Session created
          headers:
            Set-Cookie:
              schema: { type: string }
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AuthSession" }
        "400":
          description: Malformed or invalid credentials body
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "429":
          description: Login rate limit exceeded
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
...
  /api/v1/auth/me:
    get:
      tags: [auth]
      operationId: getMe
      summary: Current administrator
      security:
        - sessionAuth: []
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema: { $ref: "#/components/schemas/User" }
        "401":
          description: Not authenticated
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "403":
          description: Authenticated but not permitted
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
  /api/v1/nodes:
    get:
      tags: [nodes]
      operationId: listNodes
      summary: List nodes
      security:
        - sessionAuth: []
      parameters:
        - { $ref: "#/components/parameters/PageParam" }
        - { $ref: "#/components/parameters/PageSizeParam" }
        - { $ref: "#/components/parameters/SortParam" }
        - { $ref: "#/components/parameters/FilterParam" }
      responses:
        "200":
          description: Page of nodes
          content:
            application/json:
              schema: { $ref: "#/components/schemas/NodePage" }
    post:
      tags: [nodes]
      operationId: createNode
      summary: Create a node
      security:
        - sessionAuth: []
      parameters:
        - { $ref: "#/components/parameters/IdempotencyKeyHeader" }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/NodeCreate" }
      responses:
        "201":
          description: Node created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Node" }
        "400":
          description: Malformed request
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "409":
          description: Idempotency-key conflict or duplicate
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "413":
          description: Request body too large
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "422":
          description: Validation error
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "429":
          description: Per-principal rate limit exceeded
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "500":
          description: Unexpected server error
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
...
  /api/v1/forwards:
    get:
      tags: [forwards]
      operationId: listForwards
      summary: List forwards
      security:
        - sessionAuth: []
      parameters:
        - { $ref: "#/components/parameters/PageParam" }
        - { $ref: "#/components/parameters/PageSizeParam" }
        - { $ref: "#/components/parameters/SortParam" }
        - { $ref: "#/components/parameters/FilterParam" }
      responses:
        "200":
          description: Page of forwards
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ForwardPage" }
    post:
      tags: [forwards]
      operationId: createForward
      summary: Create a forward
      security:
        - sessionAuth: []
      parameters:
        - { $ref: "#/components/parameters/IdempotencyKeyHeader" }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ForwardCreate" }
      responses:
        "201":
          description: Forward created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Forward" }
        "409":
          description: Port/state conflict
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "422":
          description: Validation error
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
...
  /api/v1/hooks/definitions:
    get:
      tags: [hooks]
      operationId: listHookDefinitions
      summary: List hook definitions
      security:
        - sessionAuth: []
      responses:
        "200":
          description: Hook definitions
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/HookDefinition" }
    post:
      tags: [hooks]
      operationId: createHookDefinition
      security:
        - sessionAuth: []
      parameters:
        - { $ref: "#/components/parameters/IdempotencyKeyHeader" }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/HookDefinitionCreate" }
      responses:
        "201":
          description: Hook definition created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/HookDefinition" }
...
  /api/v1/hooks/secrets:
    get:
      tags: [hooks]
      operationId: listHookSecrets
      summary: List hook secret metadata (never the secret values)
      security:
        - sessionAuth: []
      responses:
        "200":
          description: Hook secret metadata
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/HookSecret" }
    post:
      tags: [hooks]
      operationId: createHookSecret
      security:
        - sessionAuth: []
      parameters:
        - { $ref: "#/components/parameters/IdempotencyKeyHeader" }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/HookSecretCreate" }
      responses:
        "201":
          description: Hook secret created (value never returned)
          content:
            application/json:
              schema: { $ref: "#/components/schemas/HookSecret" }
...
  /api/v1/audit:
    get:
      tags: [observability]
      operationId: listAudit
      summary: List admin audit log entries
      security:
        - sessionAuth: []
      parameters:
        - { $ref: "#/components/parameters/PageParam" }
        - { $ref: "#/components/parameters/PageSizeParam" }
        - { $ref: "#/components/parameters/CursorParam" }
      responses:
        "200":
          description: Page of audit entries
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AuditPage" }
  /api/v1/traffic:
    get:
      tags: [observability]
      operationId: getTraffic
      summary: Traffic rollups
      security:
        - sessionAuth: []
      parameters:
        - { $ref: "#/components/parameters/PageParam" }
        - { $ref: "#/components/parameters/PageSizeParam" }
      responses:
        "200":
          description: Traffic rollups
          content:
            application/json:
              schema: { $ref: "#/components/schemas/TrafficPage" }
  /api/v1/events:
    get:
      tags: [observability]
      operationId: streamEvents
      summary: Server-sent events with durable Last-Event-ID cursor
      security:
        - sessionAuth: []
      parameters:
        - { $ref: "#/components/parameters/LastEventIdHeader" }
      responses:
        "200":
          description: text/event-stream
          headers:
            Cache-Control:
              schema: { type: string }
          content:
            text/event-stream:
              schema:
                type: object
                properties:
                  event:
                    type: string
                  id:
                    type: string
                  data:
                    type: string
  /api/v1/settings:
    get:
      tags: [settings]
      operationId: getSettings
      summary: Read global settings (secret values redacted)
      security:
        - sessionAuth: []
      responses:
        "200":
          description: Settings
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Settings" }
    put:
      tags: [settings]
      operationId: putSettings
      summary: Replace global settings
      security:
        - sessionAuth: []
      parameters:
        - { $ref: "#/components/parameters/IfMatchHeader" }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/SettingsPatch" }
      responses:
        "200":
          description: Settings saved
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Settings" }
        "412":
          description: ETag/revision mismatch
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "428":
          description: ETag/If-Match required
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
components:
  securitySchemes:
    sessionAuth:
      type: apiKey
      in: cookie
      name: antinat_session
  parameters:
    PageParam:
      name: page
      in: query
      description: One-based page number
      schema: { type: integer, minimum: 1, default: 1 }
    PageSizeParam:
      name: page_size
      in: query
      description: Page size, 1..200
      schema: { type: integer, minimum: 1, maximum: 200, default: 50 }
    SortParam:
      name: sort
      in: query
      description: Sort field and direction, e.g. -created_at
      schema: { type: string }
    FilterParam:
      name: filter
      in: query
      description: Field filter, e.g. state=ONLINE
      schema: { type: string }
    CursorParam:
      name: cursor
      in: query
      description: Opaque pagination cursor (preferred over page for streams)
      schema: { type: string }
    IfMatchHeader:
      name: If-Match
      in: header
      required: true
      description: ETag precondition; missing returns 428
      schema: { type: string }
    IdempotencyKeyHeader:
      name: Idempotency-Key
      in: header
      required: true
      description: Durable idempotency key for creates
      schema: { type: string, minLength: 8, maxLength: 128 }
    LastEventIdHeader:
      name: Last-Event-ID
      in: header
      required: false
      description: Durable SSE event cursor; replay from this id
      schema: { type: string }
...
  schemas:
...
    Error:
      type: object
      required: [code, message, request_id]
      properties:
        code: { type: string, description: Error code from docs/error-codes.md }
        message: { type: string }
        request_id: { type: string }
        details:
          type: object
          additionalProperties: true
    LoginRequest:
      type: object
      required: [username, password]
      properties:
        username: { type: string }
        password: { type: string, format: password }
    AuthSession:
      type: object
      required: [user, session_id]
      properties:
        user: { $ref: "#/components/schemas/User" }
        session_id: { type: string }
    User:
      type: object
      required: [id, username]
      properties:
        id: { type: string }
        username: { type: string }
    Node:
      type: object
      required: [id, name, control_state, created_at, etag]
      properties:
        id: { type: string }
        name: { type: string }
        control_state: { type: string, enum: [ONLINE, OFFLINE] }
        os_arch: { type: string }
        network_fingerprint: { type: string }
        created_at: { type: string, format: date-time }
        etag: { type: string }
    NodeCreate:
      type: object
      properties:
        name: { type: string, description: Optional; generated when empty }
...
    NodePage:
      type: object
      required: [items, page, page_size, total]
      properties:
        items:
          type: array
          items: { $ref: "#/components/schemas/Node" }
        page: { type: integer }
        page_size: { type: integer }
        total: { type: integer }
        next_cursor: { type: string }
...
    Forward:
      type: object
      required: [id, node_id, name, protocol, target, strategy, desired_revision, etag]
      properties:
        id: { type: string }
        node_id: { type: string }
        name: { type: string }
        protocol: { type: string, enum: [tcp, udp] }
        target: { type: string, description: IPv4:port }
        strategy: { type: string }
        source_interface: { type: string }
        requested_local_port: { type: integer }
        requested_public_port: { type: integer }
        manual_expected_endpoint: { type: string }
        rate_limit_bps: { type: integer }
        detailed_stats: { type: boolean }
        publish_scheme: { type: string, enum: [http, https] }
        published_host: { type: string }
        custom_uri_template: { type: string }
        desired_revision: { type: integer }
        etag: { type: string }
        states:
          $ref: "#/components/schemas/ActivationStates"
    ActivationStates:
      type: object
      properties:
        control_state: { type: string }
        listener_state: { type: string }
        mapping_state: { type: string }
        keepalive_state: { type: string }
        wan_reachability_state: { type: string }
        return_path_state: { type: string }
        target_health_state: { type: string }
        publication_state: { type: string }
        data_plane_state: { type: string }
    ForwardCreate:
      type: object
      required: [node_id, name, protocol, target, strategy]
      properties:
        node_id: { type: string }
        name: { type: string }
        protocol: { type: string, enum: [tcp, udp] }
        target: { type: string }
        strategy: { type: string }
        source_interface: { type: string }
        requested_local_port: { type: integer }
        requested_public_port: { type: integer }
        manual_expected_endpoint: { type: string }
        rate_limit_bps: { type: integer }
        detailed_stats: { type: boolean }
        publish_scheme: { type: string, enum: [http, https] }
        published_host: { type: string }
        custom_uri_template: { type: string }
...
    ForwardPage:
      type: object
      required: [items, page, page_size, total]
      properties:
        items:
          type: array
          items: { $ref: "#/components/schemas/Forward" }
        page: { type: integer }
        page_size: { type: integer }
        total: { type: integer }
        next_cursor: { type: string }
...
    HookDefinition:
      type: object
      required: [id, name, kind, etag]
      properties:
        id: { type: string }
        name: { type: string }
        kind: { type: string, enum: [webhook] }
        url: { type: string, format: uri }
        etag: { type: string }
    HookDefinitionCreate:
      type: object
      required: [name, kind, url]
      properties:
        name: { type: string }
        kind: { type: string, enum: [webhook] }
        url: { type: string, format: uri }
...
    HookSecret:
      type: object
      required: [id, secret_id, algorithm, etag]
      properties:
        id: { type: string }
        secret_id: { type: string }
        algorithm: { type: string }
        etag: { type: string }
    HookSecretCreate:
      type: object
      required: [secret_id, algorithm, value]
      properties:
        secret_id: { type: string }
        algorithm: { type: string }
        value: { type: string, format: password, description: Never returned }
    AuditPage:
      type: object
      required: [items, next_cursor]
      properties:
        items:
          type: array
          items: { $ref: "#/components/schemas/AuditEntry" }
        next_cursor: { type: string }
    AuditEntry:
      type: object
      required: [id, event, created_at]
      properties:
        id: { type: string }
        event: { type: string }
        detail: { type: object, additionalProperties: true }
        created_at: { type: string, format: date-time }
    TrafficPage:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items: { $ref: "#/components/schemas/TrafficRollup" }
    TrafficRollup:
      type: object
      required: [forward_id, period, bytes_in, bytes_out]
      properties:
        forward_id: { type: string }
        period: { type: string }
        bytes_in: { type: integer }
        bytes_out: { type: integer }
        legacy_connection_count: { type: integer }
        fully_effective: { type: boolean }
    Settings:
      type: object
      required: [etag]
      properties:
        etag: { type: string }
        controller_endpoint: { type: string }
        private_site: { type: boolean }
        language: { type: string, enum: [zh, en] }
        tcp_stun_servers:
          type: array
          items: { type: string }
        udp_stun_servers:
          type: array
          items: { type: string }
        probe_provider_pin: { type: string }
        probe_vantage: { type: string }
        detection_scheduler: { type: string }
        retention_days: { type: integer }
        update_policy: { type: string }
    SettingsPatch:
      type: object
      properties:
        controller_endpoint: { type: string }
        private_site: { type: boolean }
        language: { type: string, enum: [zh, en] }
        tcp_stun_servers:
          type: array
          items: { type: string }
        udp_stun_servers:
          type: array
          items: { type: string }
        probe_provider_pin: { type: string }
        probe_vantage: { type: string }
        detection_scheduler: { type: string }
        retention_days: { type: integer }
        update_policy: { type: string }